tests/hlsl: Add a test for IsHelperLane.
Shader models before 6.6 allow the use of this intrinsic, but it does not
result in a call to IsHelperLane in the DXIL code. Code is emitted which
evaluates 'SV_Coverage == 0' and stores the result in a 32-bit scalar
global which is declared with a zero initialiser. When the global is
read, the value is truncated to bool.