vkd3d/state: Limit each virtual heap descriptor set to only one vkd3d descriptor type.
This allows Vulkan descriptor pools to contain only those descriptor
types used to implement one D3D12 descriptor range type, so pools no
longer contain a potentially large number of unused descriptors when the
pool's supply of one type is fully allocated. This decreases device
memory usage in RDR2, for example. However, it reduces the number of
descriptor sets available for unbounded ranges, which could break
unbounded range support on hardware with a small set count limit.
Unbounded ranges are unlikely to work well with virtual heaps, so this
may never prove to be an issue.