vkd3d-shader/dxil: Handle 16-bit minimum-precision registers in the DXIL parser instead of the backend.
Where the VSIR validator expects types to be of equal width, emitting
16-bit registers to represent minimum-precision registers can result in
validation failure due to mixed widths. It is much simpler to handle
these in the DXIL parser than to make the validator check the validity
of mixed widths, and this also simplifies the implementation of native
16-bit support in backends.

The SPIR-V backend currently ignores minimum-precision flags, so
setting them has no effect.