vkd3d-shader/dxil: Handle 16-bit minimum-precision indexable temp declarations in the DXIL parser instead of the backend.
For simplicity, a distinction is made between declaration instructions
which use a member of the declaration union in struct
vkd3d_shader_instruction and all other instructions. Minimum-precision
data types in declarations are handled during parsing so we do not need
to handle instruction opcodes in vsir_program_convert_min_precision().
These declarations are not referred to anywhere else, so it is safe to
convert them early.