vkd3d/device: Do not use more than a few millions descriptors for Vulkan heaps.
Currently we allocate all the descriptors that the Vulkan
implementation declares available. Recently this began being a
probem, because since Mesa 24.2.7 the Intel driver advertises
more than 200 million descriptors, but pipeline compilation takes
linear RAM in the number of used descriptors. This means that
compiling even a simple shader requires 10-20 GB of RAM.

Even of the best tier, D3D12 applications should not expect to have
more than a million usable descriptors, so there's not reason to
allocate more, even if the Vulkan implementation allows it.